Rename omtLinkProvider to omtDocumentInformationProvider
